+++ sdk/ml/azure-ai-ml/CHANGELOG.md
@@ -10,6 +10,22 @@
 
 ### Other Changes
 
+- The following classes will still be able to be imported from azure.ai.ml, but the import is deprecated and emits a warning. Instead, please import them from azure.ai.ml.entities.
+  - AmlTokenConfiguration
+  - ManagedIdentityConfiguration
+  - UserIdentityConfiguration
+- The following classes will still be able to be imported from azure.ai.ml.entities, but the import is deprecated and emits a warning. Instead, please import them from azure.ai.ml.sweep.
+  - Choice
+  - Uniform
+  - LogUniform
+  - QLogUniform
+  - QUniform
+  - QLogNormal
+  - QNormal
+  - LogNormal
+  - Normal
+  - Randint
+
 ## 1.14.0 (unreleased)
 
 ### Features Added